Rename tools module to util
Rename the tools module to util. It's shorter and the more common name
for the kind of stuff it contains.
