+++ src/dwarf/debug_info.rs
@@ -21,7 +21,7 @@
 //! It will walk through the data in the `.debug_info` and
 //! `.debug_abbrev` section to return Units.
 use super::constants::*;
-use crate::tools::{
+use crate::util::{
     decode_leb128_128, decode_leb128_128_s, decode_udword, decode_uhalf, decode_uword,
     extract_string,
 };
